+++ packages/callout/src/index.ts
@@ -1,2 +1,2 @@
 export { default } from './Callout/Callout';
-export { type CalloutProps, Variant } from './Callout/types';
+export { type CalloutProps, Variant } from './Callout/Callout.types';
